+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:4.2
+// swift-tools-version:5.0
 //
 //  Package.swift
 //  SwiftLinkPreview
@@ -11,6 +11,12 @@ import PackageDescription
 
 let package = Package(
     name: "SwiftLinkPreview",
+    platforms: [
+        .iOS("8.0"),
+        .macOS("10.11"),
+        .tvOS("9.0"),
+        .watchOS("2.0")
+    ],
     products: [
         .library(
             name: "SwiftLinkPreview",
@@ -23,5 +29,6 @@ let package = Package(
             dependencies: [],
             path: "Sources"
         ),
-    ]
+    ],
+    swiftLanguageVersions: [.v5]
 )